Update routes and remove unused code
